+++ CHANGELOG.md
@@ -5,9 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased ]
+## [Unreleased 0.2.0-beta]
 ### Added
-- Added Smartmeter and Powerflow sample data
+- Added support for realtime meter and powerflow data
+- Added smartmeter, powerflow and cummulation inverter sample data
 
 ## [0.1.0] - 2023-06-01
 ### Added